Limit Updates, Bug Fixes
Fixed a bug with the WASD controls being flipped. Fixed a bug with the boost not working as intended. Made it so the hub sends updates as soon as possible but only if the update is different from the previous state. Limited number of updates to 10 per second to keep the receiver from getting overwhelmed