Game Logic Infrastructure
Added the game logic for the hub. The game will automatically reset after the game is over, unless the score is tied, in which case it goes into overtime. Overtime is sudden death, and the game will end immediately if a goal is scored. If no goal is scored in time, the match results in a draw. Once the game is over, there is a small grace period before resetting so that the web page can display the victory. The web page still needs to be updated to display this information. Also added manual keys for resetting the game state (red score, blue score, time left, overtime, and winner) and for resetting any or all of the cars