WebPage Features
Added a catch for SerialExceptions in case a goal gets unplugged, we will just reset the hub manually if that happens but it won't crash. Added comments and changed the match times for demo day. Webpage now sends players back to the login page when the game is over. Claimed cars are now locked unless the name in the box is the same as the name claiming the car